+++ src/main/java/com/enigma/service/FoodService.java
@@ -4,7 +4,6 @@ import com.enigma.entity.FoodEntities;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface FoodService {
@@ -14,6 +13,6 @@ public interface FoodService {
     void deleteFoodById(String idFood);
     Page<FoodEntities> getAllFoodPagination(Pageable pageable);
     void deductQuantityFood(String idFood,Integer quantity);
-    BigDecimal getFoodPriceById(String idFood);
+    Integer getFoodPriceById(String idFood);
     FoodEntities updateFoodQuantity(FoodEntities newData);
 }
