+++ src/main/java/com/enigma/service/serviceImpl/ImplementFoodService.java
@@ -9,7 +9,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -54,7 +53,7 @@ public class ImplementFoodService implements FoodService {
     }
 
     @Override
-    public BigDecimal getFoodPriceById(String idFood) {
+    public Integer getFoodPriceById(String idFood) {
         return getFoodById(idFood).getPrice();
     }
